feat: add support for one line output
This changeset adds support for a new paramater output.oneline that allows a
single line per alert using the alert_fast log module. This is particulary
interesting when forwarding alerts to SIEM systems and context (payload) is
needed in a compact -and searchable- form.

Also some whitespace at the end of lines and blank lines at the end of files
are removed.